deployment/model-upload/pyproject.toml
[project]
name = "arthur-model-upload"
version = "1.0.0"
description = "Job to transfer ML models to storage for airgapped environments (S3/GCS/PVC)"
authors = [{ name = "Arthur", email = "engineering@arthur.ai" }]
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
  "huggingface-hub>=0.36.0",
]

[project.optional-dependencies]
s3  = ["boto3>=1.28.0,<2"]
gcs = ["google-cloud-storage>=2.10.0"]
k8s = ["tiktoken>=0.7.0"]

[project.scripts]
upload   = "upload_models:main"
download = "download_models:main"

[dependency-groups]
dev = [
  "pytest>=8.0.0,<9",
  "pytest-asyncio>=0.26.0",
  "black>=26.0.0",
  "mypy>=1.13.0,<2",
  "isort>=5.13.0,<6",
  "autoflake>=2.3.0,<3",
  "boto3-stubs[s3]>=1.28.0,<2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.black]
target-version = ['py312']
line-length = 88

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-vv"
testpaths = ["tests"]
asyncio_mode = "auto"
